+++ stop_words/__init__.py
@@ -1,7 +1,7 @@
 import json
 import os
 
-__VERSION__ = (2015, 2, 21)
+__VERSION__ = (2015, 2, 23)
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
 STOP_WORDS_DIR = os.path.join(CURRENT_DIR, 'stop-words')
 STOP_WORDS_CACHE = {}
@@ -11,7 +11,7 @@ with open(os.path.join(STOP_WORDS_DIR, 'languages.json'), 'rb') as map_file:
     buffer = buffer.decode('ascii')
     LANGUAGE_MAPPING = json.loads(buffer)
 
-AVAILABLE_LANGUAGES = LANGUAGE_MAPPING.values()
+AVAILABLE_LANGUAGES = list(LANGUAGE_MAPPING.values())
 
 
 def get_version():
@@ -25,7 +25,7 @@ class StopWordError(Exception):
     pass
 
 
-def get_stop_words(language):
+def get_stop_words(language, cache=True):
     """
     :type language: basestring
 
@@ -39,7 +39,7 @@ def get_stop_words(language):
                 language
             ))
 
-    if language in STOP_WORDS_CACHE:
+    if cache and language in STOP_WORDS_CACHE:
         return STOP_WORDS_CACHE[language]
 
     language_filename = os.path.join(STOP_WORDS_DIR, language + '.txt')
@@ -47,6 +47,7 @@ def get_stop_words(language):
         with open(language_filename, 'rb') as language_file:
             stop_words = [line.decode('utf-8').strip()
                           for line in language_file.readlines()]
+            stop_words = apply_filters(stop_words, language)
     except IOError:
         raise StopWordError(
             '{0}" file is unreadable, check your installation.'.format(
@@ -54,9 +55,55 @@ def get_stop_words(language):
             )
         )
 
-    STOP_WORDS_CACHE[language] = stop_words
+    if cache:
+        STOP_WORDS_CACHE[language] = stop_words
 
     return stop_words
+
+_filters = {None: []}
+
+
+def apply_filters(stopwords, language):
+    """
+    Apply registered filters to stopwords
+    :param stopwords: list
+    :param language: string
+    :return: filtered stopwords
+    """
+    if language in _filters:
+        for func in _filters[language]:
+            stopwords = func(stopwords)
+
+    for func in _filters[None]:
+        stopwords = func(stopwords, language)
+
+    return stopwords
+
+
+def add_filter(func, language=None):
+    """
+    Register filters for specific language.
+    If language == None the filter applies for all languages.
+    Filter will not apply for stop words in cache.
+    :param func: callable
+    :param language: string|None
+    :return:
+    """
+    if language not in _filters:
+        _filters[language] = []
+    _filters[language].append(func)
+
+
+def remove_filter(func, language=None):
+    """
+    :param func:
+    :param language:
+    :return:
+    """
+    if not (language in _filters and func in _filters[language]):
+        return False
+    _filters[language].remove(func)
+    return True
 
 
 def safe_get_stop_words(language):